[#28 refactoring] 리부 반영 위치 수정
