+++ data/src/main/kotlin/com/mashup/dorabangs/data/datasource/remote/api/DoraUrlCheckRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.mashup.dorabangs.data.datasource.remote.api
+
+import com.mashup.dorabangs.data.datasource.save.DoraUrlCheckResponse
+
+interface DoraUrlCheckRemoteDataSource {
+    suspend fun checkDataSource(urlLink: String): DoraUrlCheckResponse
+}
