--- data/src/main/kotlin/com/mashup/dorabangs/data/datasource/save/DoraUrlCheckRemoteDataSource.kt
@@ -1,5 +1,0 @@
-package com.mashup.dorabangs.data.datasource.save
-
-interface DoraUrlCheckRemoteDataSource {
-    suspend fun checkDataSource(urlLink: String): DoraUrlCheckResponse
-}
