+++ data/src/main/kotlin/com/mashup/dorabangs/data/datasource/save/DoraUrlCheckResponse.kt
@@ -1,8 +1,8 @@
 package com.mashup.dorabangs.data.datasource.save
 
 data class DoraUrlCheckResponse(
-    val urlLink: String?,
-    val title: String?,
-    val thumbnailUrl: String?,
-    val isShortLink: Boolean?,
+    val urlLink: String = "",
+    val title: String = "",
+    val thumbnailUrl: String = "",
+    val isShortLink: Boolean = false,
 )
